+++ bindings/python/docs/source/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 Changes in Version 0.3.0
 ------------------------
 - Support for `ObjectId` `bson` type.
+- Improve error message when schema contains an unsupported type.
 
 Changes in Version 0.2.0
 ------------------------